refactor(main-api): 이미지 저장 구현 main-api로 이동
- 이미지가 domain에 존재하면 안 됨
- 이미지 저장 구현이 jpa 기술이 아니므로 domain-jpa에 존재하는 것이 적절치 않음
