+++ modules/main-api/src/main/java/com/whoz_in/main_api/shared/utils/ImageUploadStorage.java
@@ -1,4 +1,4 @@
-package com.whoz_in.domain.member;
+package com.whoz_in.main_api.shared.utils;
 
 import com.whoz_in.domain.member.model.ImageId;
